requirements.txt
pytest==8.4.2
requests==2.32.5
allure-pytest==2.15.0
pytest-playwright==0.7.1
python-dotenv==1.1.1
jsonschema==4.25.1
allure-pytest==2.15.0